requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
abseil-cpp=20211102.0=h0e60522_0
absl-py=1.4.0=pyhd8ed1ab_0
aiohttp=3.8.5=py38h2bbff1b_0
aiosignal=1.3.1=pyhd8ed1ab_0
appdirs=1.4.4=pyh9f0ad1d_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=py38haa95532_0
attrs=23.1.0=pyh71513ae_1
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.6.2=pyhd8ed1ab_0
brotli=1.0.9=h0e60522_4
brotlipy=0.7.0=py38h294d835_1004
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2023.7.22=h56e8100_0
cachetools=5.3.1=pyhd8ed1ab_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py38hd8c33c5_0
charset-normalizer=3.2.0=pyhd8ed1ab_0
click=8.1.7=win_pyh7428d3b_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.0.4=py38hbd9d945_0
cryptography=3.4.8=py38hb7941b4_1
cudatoolkit=11.0.221=h74a9793_0
cycler=0.11.0=pyhd8ed1ab_0
decorator=5.1.1=pyhd3eb1b0_0
executing=0.8.3=pyhd3eb1b0_0
fonttools=4.37.1=py38h294d835_0
freetype=2.12.1=ha860e81_0
frozenlist=1.3.3=py38h2bbff1b_0
google-auth=2.23.0=pyh1a96a4e_0
google-auth-oauthlib=1.0.0=pyhd8ed1ab_1
grpc-cpp=1.48.2=hf108199_0
grpcio=1.48.2=py38hf108199_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.8.0=pyha770c72_0
intel-openmp=2023.1.0=h59b6b97_46319
ipython=8.8.0=py38haa95532_0
jedi=0.18.1=py38haa95532_1
jpeg=9b=hb83a4c4_2
jsonpatch=1.32=pyhd8ed1ab_0
jsonpointer=2.4=py38haa244fe_2
kiwisolver=1.4.4=py38hbd9d945_0
lerc=3.0=hd77b12b_0
libdeflate=1.17=h2bbff1b_0
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libtiff=4.5.0=h8a3f274_0
libuv=1.44.2=h2bbff1b_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_1
lz4-c=1.9.4=h2bbff1b_0
markdown=3.4.4=pyhd8ed1ab_0
matplotlib=3.6.2=py38haa244fe_0
matplotlib-base=3.6.2=py38h1094b8e_0
matplotlib-inline=0.1.6=py38haa95532_0
mkl=2023.1.0=h6b88ed4_46357
mkl-service=2.4.0=py38h2bbff1b_1
mkl_fft=1.3.8=py38h2bbff1b_0
mkl_random=1.2.4=py38h59b6b97_0
multidict=6.0.2=py38h2bbff1b_0
munkres=1.1.4=pyh9f0ad1d_0
ninja=1.10.2=haa95532_5
ninja-base=1.10.2=h6d14046_5
numpy=1.24.3=py38h79a8e48_1
numpy-base=1.24.3=py38h8a87ada_1
oauthlib=3.2.2=pyhd8ed1ab_0
openssl=1.1.1w=h2bbff1b_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.3=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.3.0=py38hdc2b20a_1
pip=23.2.1=py38haa95532_0
pooch=0.6.0=py38_0
prompt-toolkit=3.0.36=py38haa95532_0
protobuf=3.20.3=py38haa244fe_1
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.5.0=pyhd8ed1ab_0
pyasn1-modules=0.3.0=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.8.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=3.1.1=pyhd8ed1ab_0
pyqt=5.9.2=py38hd77b12b_6
pysocks=1.7.1=pyh0701188_6
python=3.8.13=h6244533_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.7.1=py3.8_cuda110_cudnn8_0
pytz=2023.3.post1=pypi_0
pyu2f=0.1.5=pyhd8ed1ab_0
qt=5.9.7=vc14h73c81de_0
re2=2022.04.01=h0e60522_0
requests=2.31.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scipy=1.10.1=py38hdcfc7df_1
setuptools=68.0.0=py38haa95532_0
sip=4.19.13=py38hd77b12b_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
tbb=2021.8.0=h59b6b97_0
tensorboard=2.14.0=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py38haa244fe_0
tk=8.6.12=h2bbff1b_0
torch-tb-profiler=0.4.1=pypi_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu110
tornado=6.2=py38h294d835_0
tqdm=4.66.1=pyhd8ed1ab_0
traitlets=5.7.1=py38haa95532_0
typing_extensions=4.7.1=py38haa95532_0
tzdata=2023.3=pypi_0
unicodedata2=14.0.0=py38h294d835_1
urllib3=1.26.15=pyhd8ed1ab_0
vc=14.2=h21ff451_1
visdom=0.2.4=pyhd8ed1ab_0
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
websocket-client=1.6.3=pyhd8ed1ab_0
werkzeug=2.1.2=pyhd8ed1ab_1
wheel=0.38.4=py38haa95532_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
xz=5.4.2=h8cc25b3_0
yarl=1.7.2=py38h294d835_2
zipp=3.16.2=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
